Remove trainer_util and integrate functions to trainer_main and engine_utils
