+++ src/netspresso_trainer/__init__.py
@@ -2,7 +2,6 @@ from pathlib import Path
 
 from netspresso_trainer.evaluator_main import evaluation_cli
 from netspresso_trainer.trainer_main import parse_args_netspresso, train_cli, train_with_yaml
-from netspresso_trainer.trainer_util import validate_config
 
 ### Starting from v0.0.9, the default train function runs with yaml configuration
 train = train_with_yaml
