+++ src/netspresso_trainer/trainer_main.py
@@ -1,15 +1,94 @@
 import os
+import subprocess
 from pathlib import Path
 from typing import List, Literal, Optional, Union
 
-from netspresso_trainer.trainer_common import train_common
-from netspresso_trainer.trainer_util import (
-    set_arguments,
-    train_with_yaml_impl,
-)
-from .utils.engine_utils import parse_args_netspresso, parse_gpu_ids
+import torch
+from omegaconf import OmegaConf
 
-LOG_LEVEL = os.getenv('LOG_LEVEL', 'INFO')
+from netspresso_trainer.trainer_common import train_common
+from netspresso_trainer.utils.engine_utils import (
+    parse_args_netspresso,
+    parse_gpu_ids,
+    set_arguments,
+    get_gpus_from_parser_and_config,
+    validate_train_config,
+)
+from netspresso_trainer.utils.engine_utils import LOG_LEVEL
+
+
+def run_distributed_training_script(gpu_ids, data, augmentation, model, training, logging, environment, log_level,
+                                    task, model_name, is_graphmodule_training, logging_dir):
+
+    command = [
+        "--data", data,
+        "--augmentation", augmentation,
+        "--model", model,
+        "--training", training,
+        "--logging", logging,
+        "--environment", environment,
+        "--log-level", log_level,
+        "--task", task,
+        "--model-name", model_name,
+        "--is-graphmodule-training", is_graphmodule_training,
+        "--logging-dir", logging_dir,
+    ]
+
+    # Distributed training script
+    command = [
+        'python', '-m', 'torch.distributed.launch',
+        f'--nproc_per_node={len(gpu_ids)}',  # GPU #
+        f"{Path(__file__).absolute().parent / 'trainer_main.py'}", *map(str, command)
+    ]
+
+    # Run subprocess
+    process = subprocess.Popen(command)
+
+    try:
+        process.wait()
+    except KeyboardInterrupt:
+        print("Interrupted. Terminating the training process...")
+        process.terminate()
+        process.wait()
+
+
+def train_with_yaml_impl(gpus: Optional[Union[List, int]], data: Union[Path, str], augmentation: Union[Path, str],
+                         model: Union[Path, str], training: Union[Path, str],
+                         logging: Union[Path, str], environment: Union[Path, str], log_level: str = LOG_LEVEL):
+    conf_environment = OmegaConf.load(environment).environment
+    gpus = get_gpus_from_parser_and_config(gpus, conf_environment)
+    assert isinstance(gpus, (list, int))
+
+    gpu_ids_str = ','.join(map(str, gpus)) if isinstance(gpus, list) else str(gpus)
+    os.environ['CUDA_VISIBLE_DEVICES'] = gpu_ids_str
+    torch.cuda.empty_cache()  # Reinitialize CUDA to apply the change
+
+    conf = set_arguments(data=data,
+                         augmentation=augmentation,
+                         model=model,
+                         training=training,
+                         logging=logging,
+                         environment=environment)
+    config_summary = validate_train_config(conf)
+
+    try:
+        if isinstance(gpus, int):
+            train_common(
+                conf,
+                task=config_summary.task,
+                model_name=config_summary.model_name,
+                is_graphmodule_training=config_summary.is_graphmodule_training,
+                logging_dir=config_summary.logging_dir,
+                log_level=log_level
+            )
+        else:
+            run_distributed_training_script(
+                gpus, data, augmentation, model, training, logging, environment, log_level,
+                config_summary.task, config_summary.model_name, config_summary.is_graphmodule_training, config_summary.logging_dir
+            )
+        return config_summary.logging_dir
+    except Exception as e:
+        raise e
 
 
 def train_with_yaml(
